+++ src/main/java/uk/org/ukfederation/mda/dom/StringXPathValidationStage.java
@@ -1,0 +1,34 @@
+/*
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ *    http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+package uk.org.ukfederation.mda.dom;
+
+import javax.annotation.Nonnull;
+
+import org.w3c.dom.Node;
+
+import net.shibboleth.metadata.pipeline.Stage;
+
+/**
+ * A {@link Stage} allowing validation of the results of an XPath expression
+ * as a series of {@link String} values.
+ */
+public class StringXPathValidationStage extends AbstractXPathValidationStage<String> {
+
+    @Override
+    protected String convert(@Nonnull final Node node) {
+        return node.getTextContent();
+    }
+
+
+}